Fix compiler warnings related to int vs size_t

+++ src/liboslexec/shadingsys.cpp
@@ -1308,15 +1308,15 @@ ShadingSystemImpl::getattribute (ShaderGroup *group, string_view name,
         return true;
     }
     if (name == "num_entry_layers" && type.basetype == TypeDesc::INT) {
-        size_t n = 0;
-        for (size_t i = 0;  i < group->nlayers();  ++i)
+        int n = 0;
+        for (int i = 0;  i < group->nlayers();  ++i)
             n += group->layer(i)->entry_layer();
         *(int *)val = n;
         return true;
     }
     if (name == "entry_layers" && type.basetype == TypeDesc::STRING) {
         size_t n = 0;
-        for (size_t i = 0;  i < group->nlayers() && i < type.numelements();  ++i)
+        for (size_t i = 0;  i < (size_t)group->nlayers() && i < type.numelements();  ++i)
             if (group->layer(i)->entry_layer())
                 ((ustring *)val)[n++] = (*group)[i]->layername();
         for (size_t i = n;  i < type.numelements();  ++i)
@@ -2117,7 +2117,7 @@ ShadingSystemImpl::ShaderGroupBegin (string_view groupname,
             }
             // Zero-pad if we parsed fewer values than we needed
             intvals.resize (type.numelements()*type.aggregate, 0);
-            ASSERT (type.numelements()*type.aggregate == int(intvals.size()));
+            ASSERT (int(type.numelements())*type.aggregate == int(intvals.size()));
             Parameter (paramname, type, &intvals[0], lockgeom);
         } else if (type.basetype == TypeDesc::FLOAT) {
             floatvals.clear ();
@@ -2137,7 +2137,7 @@ ShadingSystemImpl::ShaderGroupBegin (string_view groupname,
             }
             // Zero-pad if we parsed fewer values than we needed
             floatvals.resize (type.numelements()*type.aggregate, 0);
-            ASSERT (type.numelements()*type.aggregate == int(floatvals.size()));
+            ASSERT (int(type.numelements())*type.aggregate == int(floatvals.size()));
             Parameter (paramname, type, &floatvals[0], lockgeom);
         } else if (type.basetype == TypeDesc::STRING) {
             stringvals.clear ();
@@ -2167,7 +2167,7 @@ ShadingSystemImpl::ShaderGroupBegin (string_view groupname,
             }
             // Zero-pad if we parsed fewer values than we needed
             stringvals.resize (type.numelements()*type.aggregate, ustring());
-            ASSERT (type.numelements()*type.aggregate == int(stringvals.size()));
+            ASSERT (int(type.numelements())*type.aggregate == int(stringvals.size()));
             Parameter (paramname, type, &stringvals[0], lockgeom);
         }
